Maybe fix changelog anchors
Sphinx didn't like that the headers included repeated link text and would
sometimes just not generate a proper anchor in that case. For reasons!

I am rapidly coming to an understanding that Sphinx is maybe okay for
documentation, but it is an _awful_ website generator. And, bluntly,
I don't trust that people use any of the other output modes enough to
make them likely any better...