Upload the SDK to S3 before doing UPM
The UPM build attempts to checkout every file that exists in main from the
branch selected for the build (e.g. 4.1). I'm not sure why it's doing that,
but it breaks the UPM build when a later SDK version has added new files,
which has happened.

The goal here is to get 4.1.16 out, and I don't believe anyone who needs
it is using UPM, so I'm okay skipping UPM for this.