Update to Xcode 14 and fix install rbenv through homebrew
`brew upgrade ruby-build` is consistently exiting with a status code of 1.
I cannot, for the life of me, figure out why. The command does not log any errors.

Initial theories were that it was related to the deprecated homebrew/cask-versions cask,
however uninstally all formula using that cask and forcibly untapping the cask did not resolve
the issue.

Testing shows that everything proceeds fine after the brew upgrade command, so I added
an explicit check for an exit code of 1 and suppress it to allow the build to continue.