Cut running the ./setup script
This script only installed awscli, which we don't use on the macos part of the
CI build. Again this was a homebrew issue -- with the build now resolved running
this would take several minutes when working with a restored cache because we
cached the dependency updates that we were trying to skip when running the setup script
initially, resulting in Homebrew, among other things, building cmake from source.

Rather than try and continue to diagnose that, we don't need the awscli for any reason
in this build step, so... just don't install it. Probably solved. We didn't even have
AWS auth here.